restructure main.js
