+++ helper/listener.js
@@ -1,0 +1,53 @@
+const path = require('path');
+const url = require('url');
+
+const exec = require('child_process').exec;
+
+const sendCustomStyles = require('./styles');
+const { downloadUpdate, sendVersionInfo } = require('./update');
+
+function activateScreenSleepListener(ipcMain) {
+  ipcMain.on('screenSleep', () => {
+    exec('xset dpms force standby');
+  });
+
+  ipcMain.on('screenWakeup', () => {
+    exec('xset s off');
+    exec('xset -dpms');
+    exec('xset s noblank');
+  });
+}
+
+function activateReloadListener(ipcMain, window) {
+  ipcMain.on('reload', () => {
+    window.loadURL(
+      url.format({
+        pathname: path.join(__dirname, 'dist/index.html'),
+        protocol: 'file:',
+        slashes: true,
+      }),
+    );
+  });
+}
+
+function activateAppInfoListener(ipcMain, window, app) {
+  ipcMain.on('appInfo', () => {
+    sendCustomStyles(window);
+    sendVersionInfo(window, app);
+  });
+}
+
+function activateUpdateListener(ipcMain, window) {
+  ipcMain.on('update', (_, updateInfo) => {
+    downloadUpdate(updateInfo, window);
+  });
+}
+
+function activateListeners(ipcMain, window, app) {
+  activateAppInfoListener(ipcMain, window, app);
+  activateScreenSleepListener(ipcMain);
+  activateReloadListener(ipcMain, window);
+  activateUpdateListener(ipcMain, window);
+}
+
+module.exports = activateListeners;
